requirements.txt
Flask==1.0.2
google-cloud-datastore==1.7.3
darksky_weather==1.8.0